requirements.txt
streamlit
google-generativeai
google-cloud-storage
Pillow
tenacity>=8.2.2 